Remove hardcoded version file lib/puppet/version.rb
This commit removes the file from lib/puppet/version.rb, which
contains a hard-coded value of the last version of puppet to be
released. This removes the last manual task and release commit
from releasing a build of puppet. In its place, the packaging
repo now has a task, version_file, that will lay this file down
from a template, with the current version that the source is at.
This task can be called pre-release so that tars/packages ship
with a version file, and can be called by devs to lay down the
file for programmatic purposes if desired.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>