(#16746) Address man page conflict with Ubuntu package
When Ubuntu packaged 2.7.11, they chose packages for the various manpages
that we shipped in source but didn't actually package ourselves. Later,
when we did package the manpages, we did so without consulting what Ubuntu
had already done. Now when upgrading from their 2.7.11 to our 3.0.0, man
page conflicts prevent successful upgrades. This commit adds a conflicts
notice to the puppet package, which conflicts with puppet-common from ubuntu.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>