(PUP-6483) Return false for nil passwords in Windows user provider
This commit builds on a previous commit by Iristyle for PUP-6483 in which
running `puppet resource user` caused login attempts a Windows sytem for each
user.

For property value retrieval, Puppet calls the getter for each property
for each resource. Previously, the password getter would ultimately call
logonUserW even with an empty/nil password. Because Windows does not allow
blank passwords, this would throw, get rescued, and return false, eventually
surfacing as a nil return value from the getter, with the side-effect of having
issued a failed logon attempt. Now the getter will return nil before issuing
the logon attempt if asked to validate an empty/nil password - it will
explicitly skip calling logonUser for password validation.

Since the password query should no longer issue a logon attempt if the password
property is nil for the resource, we add a test for this. This doesn't actually
go into the self.instances block just because self.instances actually doesn't
result in any calls to the password getter - that's done separately when
retrieving values for all properties for each provider instance.

A quick note - the previous implementation of a fix raised an exception if the
password_is? function was passed an empty or nil password. This commit modifies
this to return false and issue a warning to keep with the previous behavior
which was to return false.

New warnings are issued by the provider if a password setting attempt is made
with a nil value or empty string. They are different because as of PUP-5271 it
has not been supported to set a nil password - such an attempt is silently
ignored by Puppet::Util::Windows::ADSI::User#password=. This commit adds a new
warning for that case, as well as a recommendation not to set blank passwords.