(maint) Add test for env_windows_installdir fact
(PA-466) This commit adds a test to ensure the fact env_windows_installdir is
present and set to the correct directory. This fact is set via the
environment.bat file on Windows installations and recently was impacted by a
regression in that file, so a test is warranted. This test is added as its own
file as opposed to the 4654_facts_face.rb file because that test seemed in
spirit to be testing the functionatlity of `puppet facts` rather than any
specific fact, while this test is for an explicit fact.

Signed-off-by: Moses Mendoza <mendoza.moses@gmail.com>