(PUP-6586) Windows passwords can be blank
In doing work for PUP-6586, it was discovered that various error codes returned
via LogonUser can be used to distinguish between a failed authentication
attempt due to some account constraint or status and failure due to an
incorrect password (a notable exception being ERROR_ACCOUNT_LOCKED_OUT, 1909,
which is always returned in that circumstance regardless of password
correctness). With this information, it was determined that we can indeed
ascertain a successful logon even with a password set to a zero-length string,
which means puppet can manage a password set to a zero length string.

This commit reverses those portions of commit 70426ea which restrict
setting/reading a password set to '', and associated warning messages, allowing
puppet to continue to manage blank passwords on Windows in combination with the
previous commits associated with PUP-6586.

Of interest is that in a Windows environments with no password length
requirements (where logon with a blank password is allowed) with a user whose
password is set to '', it is possible to perform a successful LogonUserW call
with the value for password set to '' - however, if we supply `nil` instead of
'' (which we supply as FFI::Pointer::NULL), LogonuserW will also succeed -
implying that it does not distinguish between a zero-length string and NULL. In
any case, this means that, strictly speaking, we can supply nil to a logon
attempt, so we remove code that raises an exception in this case.