(maint) remove inaccurate comment in configurer.rb
Once upon a time Configurer#apply_catalog would call a method to retrieve the
catalog if needed - this is no longer the case, as the method assumes a catalog
is passed, but prior to this commit the comment still claimed a retrieval was
sometimes performed. Without the comment, it's self-explanatory what the method
does.