Partial Revert "(PUP-6659) Remove Stacktraces in API responses"
This partially reverts commit 7a765b70bb392f1d6ce9c230686f780d249720d0.

Commit 7a765b70bb392f1d6ce9c230686f780d249720d0 both removed the stacktrace
array from the http error object and introduced the stacktrace to the puppet
error log. We revert the removal of the stacktrace from the http error object
because it was a backwards-incompatible change. However, we retain the addition
of the stacktrace to the puppet error log, a) to avoid introducing _more_
backwards incompatible change and b) because in a subsequent commit we will be
replacing the stack trace string content in the http error object with a
deprecation warning and retaining the content for the logs will allow users to
still find this information if they need it.

Signed-off-by: Moses Mendoza <moses@puppet.com>