(PUP-5756) remove dependency on win32/eventlog
Remove dependency on eventlog by including our own ffi definitions of the
required functions for puppet logging. Most simple implementation - write-only
support for event log.

API compatible with previous implementation.

Uses W (unicode) versions of API functions. The spec for these functions can be
found here:
https://msdn.microsoft.com/en-us/library/windows/desktop/aa363654(v=vs.85).aspx

Replace existing log destination of win32/eventlog with
Puppet::Util::Windows::EventLog. This means we don't need to check for the
eventlog library any more as part of the eventlog feature. We keep the eventlog
feature for backwards compatibility for now, but it can be removed.

Previous implementation with Win32::EventLog actually (unnecessarily) created
two handles to the event log - one opened with source 'Application' which we
never wrote to, and a second opened with source 'Puppet', which we used. Now we
just create one handle, on initialization, with event source 'Puppet' (hence no
longer needing to supply this as a param in #report_event)