(PUP-7021) Add UTF-8-aware Etc wrapper
The Ruby Etc module will return various Struct objects containing strings in
the system's locale, or if they are invalid in the current locale, as
ASCII-8BIT (Binary). This leads to inconsistent and problematic behavior when
Puppet uses the Etc module in its providers, particularly when trying to
concatenate or compare strings which now may have different, and often
incompatible, encodings. The nameservice provider in particular makes heavy use
of Etc, which has resulted in numerous bugs reported regarding UTF-8.

Initial attempts to resolve the issues focused on the locations where string
comparison or concatenation was happening, but it became quickly clear that
there are too many places to do this, and the better solution would likely be
to control the problem where the strings first enter the system - at the Etc
module.

This commit adds a Puppet wrapper around ruby's Etc module that leverages
puppet/util/character_encoding to convert the values returned by Etc to UTF-8.
In order to avoid ever letting the original Etc values get into the system (and
with a side benefit of avoiding extraneous object creation) the Structs are
modified in-place.

It is true that this commit may introduce different failure scenarios, as
unavoidable failure will now likely be located in the string conversion process
(ie if we have truly UTF-8 incompatible strings), rather than higher in the
stack at comparison or concatenation for display, but it seems these were going
to fail anyway.

Signed-off-by: Moses Mendoza <moses@puppet.com>