(PUP-5756) do not rely on puppet in puppet/util/windows/eventlog
In order to fully remove our dependency on win32-eventlog, we need to cease
using it in ext/windows/service/daemon.rb, which containst the code that runs
Puppet as a service on Windows. This commit updates
puppet/util/windows/eventlog so that it can be used independently of Puppet.
This comes at the cost of some duplication - especially with respect to
constants and FFI utilities defined elsewhere. However, this is offset by
allowing us to centralize all of our event reporting code in one place, instead
of duplicating this in eventlog.rb.