(PUP-5756) remove dependency on win32/eventlog in service/daemon.rb
This commit is the last in a series that removes a dependency in win32/eventlog
in ext/windows/service/daemon.rb. Instead of using the gem-provided functions
for logging to eventlog, we use those provided in
puppet/util/windows/eventlog.rb. We no longer need to pass the `:source`
param to #report_event (as this is 'Puppet' by default) so this key is
removed.

Signed-off-by: Moses Mendoza <moses@puppet.com>