(PUP-7021) make utf-8 conversion destructive for perf
In scenarios where many strings are converted to UTF-8, i.e., with large groups
files containing groups with many users, the many duplicate String objects
created by the Puppet::Util::CharacterEncoding module resulted in a non-trivial
performance impact. This commit modifies the primary (only) public method,
::convert_to_utf_8, to modify the string in place, rather than creating
returning a copy. An additional optimization is made when checking if a string
would be valid UTF-8. Because the method is now destructive, ::convert_to_utf_8
is renamed to ::convert_to_utf_8!. This would be an API break, if not for the
fact that this code is unreleased - so let's get it right now.

Signed-off-by: Moses Mendoza <moses@puppet.com>