(PUP-5756) always use FFI.errno in eventlog
Prior to this commit, if the eventlog utility was used outside the context of
Puppet, the exception raised on failure to open the eventlog or report an event
was a generic RuntimeError. This commit updates this to instead define
EventlogError as either Puppet::Util::Windows::Error (which has format_error
support) if available, or SystemCallError if not, both of which accept an error
code in their initializer. Ruby's SystemCallError includes rudimentary common
error message mappings for system error codes. The raise locations are updated
to pass FFI.errno. In this way we capture useful debugging information on error
regardless of if we're leveraged from Puppet.

Signed-off-by: Moses Mendoza <moses@puppet.com>