(PUP-5756) (PUP-7021) better handling of BINARY string conversion
Prior to this commit, we would transcode to UTF-8 a string not composed of
valid UTF-8 bytes but valid in its current encoding. This presents a challenge
for BINARY strings, which if originating in an encoding other to UTF-8 are
otherwise valid, but cannot be transcoded (because we have no idea where they
came from). In order to transcode successfully, we need an originating
encoding to work with. Rather than just blindly call encode! with binary
strings, we can take one more guess as to their origin - which is that they
originated in `Encoding.default_external`. I.e., if a system is running in
Encoding::Windows_31J, and a string is returned as BINARY, we can guess that
the string might have been Encoding::Windows_31J. This perhaps substantially
increases our odds that we'll be able to conver the (binary) string to UTF-8
successfully.

Signed-off-by: Moses Mendoza <moses@puppet.com>