(PUP-7021) Ensure we don't modify string in CharacterEncoding
Prior to this commit, we were modifying the supplied string in
Puppet::Util::CharacterEncoding as a side-effect of the recent change that
added `force_encoding(Encoding.default_external)` on binary strings. In the
case where we fail to encode! that change leaves the original string with a
different external encoding than it started. This commit addresses that by
ensuring any on failure to convert we set the string back to its original
external encoding.

A quirk was accounted for in that we don't set external_encoding if the default
external encoding is UTF-8. The reason is we've already determined that this
string doesn't represent valid UTF-8 bytes so doing a force_encoding to UTF-8
doesn't get us any closer to being able to transcode it to UTF-8, and ruby
won't do anything on `encode!` to string already has as its
default_external set to the encoding supplied to encode!.

Finally, this PR cleans up/augments tests in etc_spec to account for the new
possibilities, in addition to a typo in the bytes representing a specific
character. Prior to this commit, a test string "\xfd\xf1" in EUC KR was
represented as [254, 241]. [254, 241] is actually "\xfe\xf1", a different EUC
KR character than the one we were intending to test against (though since we
made this error consistently, it was effectively the same).

Signed-off-by: Moses Mendoza <moses@puppet.com>